Fix Issue #153 : Added an option to overlay or underlay the Grid (#162)
Adds an option to overlay or underlay the Grid. Uses `showGridPosition` prop on the `ChartContainer` to control layering. `showGridPosition="over"` is the default, and draws the grid over the charts. This is useful for overlaying grids on top of area charts and is the current behavior. Alternatively `showGridPosition="under"` will draw the grid under all other elements.