Fixes #261
The previous logic added a 0 width for hidden YAxis into the widths being passed to the ChartRow. This was then used to exclude the axis from the layout. However, if there was more than one row the width passed in would be derived from axes in the other rows so this logic would fail.

This adds an additional check for the visible prop set on the YAxis which excludes it from the map of ids to YAxis. That exclusion is then used to omit the axis itself from the render but still keep its scale etc.