Added style prop to LabelAxis
LabelAxis can now receive a style prop. Applied the concept from the YAxis component to allow it to receive a style prop and merge with the default styles already present in the LabelAxis component. Only added support with styling the items that already had style objects inside the component. (Made this edit in the Github browser editor - haven't verified it, but it should get the ball rolling on adding styles to LabelAxis)