use correct form of infoStyle in defaultProps
EventMarker is currently not passing in the correct form to the Label component.  The latter expects an object which includes a `labelStyle` and a `boxStyle` (see https://github.com/esnet/react-timeseries-charts/blob/master/src/components/Label.js#L56).  The infoStyle that I've added reflects the current styling of labels in EventMarker and does not correspond to what is currently seen in the docs.