Adds a cut down version of the Baseline chart
So it can render a chart now, at least one that uses the YAxis scales ok. I had to remove the animation stuff though for now, it didn't really seem to work very well with the functional component/hooks I've been leaning towards. So I guess we add that to the TODO list and revisit it.

Also did a little reorganizing.